requirements.txt
# Hybrid Cloud Security Framework - Dependencies
# Core Framework Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication and Security
cryptography==41.0.8
pyjwt==2.8.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
saml2==5.0.0

# Machine Learning and Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.1

# Cloud Integration
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
google-cloud-storage==2.10.0
kubernetes==28.1.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2
elasticsearch==8.11.0

# API and Web Framework
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
starlette==0.27.0

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
sphinx==7.2.6

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
pydantic-settings==2.1.0
